requirements.txt
streamlit==1.41.0
streamlit-chat
pandas
numpy
boto3
langchain_aws
langchain
langchain_community
langgraph
langchain_experimental
langgraph-supervisor
langgraph-swarm
tavily-python==0.5.0
yfinance==0.2.52
rizaio==0.8.0
pytz==2024.2
beautifulsoup4==4.12.3
plotly_express==0.4.1
matplotlib==3.10.0
PyPDF2==3.0.1
opensearch-py
mcp
langchain-mcp-adapters==0.0.9
wikipedia
aioboto3
requests
uv
kaleido
diagrams
graphviz
sarif-om==1.0.4
arxiv==2.2.0
chembl-webresource-client==0.10.9
pytrials==1.0.0
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
reportlab

bedrock-agentcore
bedrock-agentcore-starter-toolkit